feat: rework LevelGeneratorBehaviour.cs into LevelManager which is subscribed to new event system
