+++ Assets/Scripts/Level/LevelGeneration/Renderer/LevelRenderer.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using DungeonDraws.Level;
 using DungeonDraws.Scripts.Systems.LevelGeneration.Plotters;
 using UnityEngine;
 
@@ -6,27 +7,29 @@ namespace DungeonDraws.Scripts.Systems.LevelGeneration.Renderer
 {
     public class LevelRenderer : ScriptableObject
     {
-        private LevelGeneratorBehaviour _behavior;
+        private LevelManager _manager;
+        private GameObject _boardHolder;
 
-        private LevelRenderer(LevelGeneratorBehaviour behaviour)
+        private LevelRenderer(LevelManager manager, GameObject boardHolder)
         {
-            _behavior = behaviour;
+            _manager = manager;
+            _boardHolder = boardHolder;
         }
 
         public void convertToMeshes(int[,] map)
         {
-            Destroy(GameObject.Find("BoardHolder"));
 
-            GameObject boardHolder = new GameObject("BoardHolder");
-            boardHolder.transform.parent = _behavior.transform;
+            Destroy(_boardHolder.transform.GetChild(0));
+            GameObject tileParent = new GameObject("TileParent");
+            tileParent.transform.parent = _manager.transform;
 
-            addMainMashes(boardHolder, map);
+            addMainMashes(tileParent, map);
             //overlapWallSerators(boardHolder, map);
         }
 
         private void addMainMashes(GameObject boardHolder, int[,] map)
         {
-            Vector3 floorSize = _behavior._floorPrefab.GetComponentInChildren<MeshRenderer>().bounds.size;
+            Vector3 floorSize = _manager.levelStyle._floorPrefab.GetComponentInChildren<MeshRenderer>().bounds.size;
             float floorSpan = floorSize.x;
             float halfFloorSpan = floorSpan * 0.5f;
 
@@ -38,62 +41,62 @@ namespace DungeonDraws.Scripts.Systems.LevelGeneration.Renderer
                     float zPos = row * floorSpan + halfFloorSpan;
 
                     if (type == DetailedTileType.Floor) {
-                        GameObject prefab = _behavior._floorPrefab;
+                        GameObject prefab = _manager.levelStyle._floorPrefab;
                         float yRot = 0f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Wall_N) {
-                        GameObject prefab = _behavior._wallPrefab;
+                        GameObject prefab = _manager.levelStyle._wallPrefab;
                         // zPos += floorSpan;
                         float yRot = 0f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Wall_E) {
-                        GameObject prefab = _behavior._wallPrefab;
+                        GameObject prefab = _manager.levelStyle._wallPrefab;
                         // xPos += floorSpan;
                         // Wall cell need to be shifted to cover floor
                         float yRot = 270f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Wall_S) {
-                        GameObject prefab = _behavior._wallPrefab;
+                        GameObject prefab = _manager.levelStyle._wallPrefab;
                         // Wall cell need to be shifted to cover floor
                         float yRot = 180f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Wall_W) {
-                        GameObject prefab = _behavior._wallPrefab;
+                        GameObject prefab = _manager.levelStyle._wallPrefab;
                         // Wall cell need to be shifted to cover floor
                         float yRot = 90f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_INN_NW) {
-                        GameObject prefab = _behavior._cornerInPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerInPrefab;
                         float yRot = 180f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_INN_NE) {
-                        GameObject prefab = _behavior._cornerInPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerInPrefab;
                         // xPos -= floorSpan;
                         // zPos += floorSpan;
                         float yRot = 270f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_INN_SE) {
-                        GameObject prefab = _behavior._cornerInPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerInPrefab;
                         float yRot = 0f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_INN_SW) {
-                        GameObject prefab = _behavior._cornerInPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerInPrefab;
                         float yRot = 90f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_OUT_NW) {
-                        GameObject prefab = _behavior._cornerOutPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerOutPrefab;
                         float yRot = 180f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_OUT_NE) {
-                        GameObject prefab = _behavior._cornerOutPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerOutPrefab;
                         float yRot = 90f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_OUT_SW) {
-                        GameObject prefab = _behavior._cornerOutPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerOutPrefab;
                         float yRot = 270f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     } else if (type == DetailedTileType.Corner_OUT_SE) {
-                        GameObject prefab = _behavior._cornerOutPrefab;
+                        GameObject prefab = _manager.levelStyle._cornerOutPrefab;
                         float yRot = 0f;
                         _instantiate(prefab, xPos, zPos, yRot, boardHolder, row, col, type.ToString());
                     }
@@ -101,23 +104,23 @@ namespace DungeonDraws.Scripts.Systems.LevelGeneration.Renderer
             }
         }
 
-        public static LevelRenderer newInstance(LevelGeneratorBehaviour behaviour)
+        public static LevelRenderer newInstance(LevelManager levelManager)
         {
             LevelRenderer renderer = ScriptableObject.CreateInstance<LevelRenderer>();
-            renderer.setBehaviour(behaviour);
+            renderer.setManager(levelManager);
             return renderer;
         }
 
-        private void setBehaviour(LevelGeneratorBehaviour dungeonGeneratorBehaviour)
+        private void setManager(LevelManager levelManager)
         {
-            _behavior = dungeonGeneratorBehaviour;
+            _manager = levelManager;
         }
 
         private void overlapWallSerators(GameObject boardHolder, int[,] map)
         {
-            GameObject prefab = _behavior._wallSeparatorPrefab;
+            GameObject prefab = _manager.levelStyle._wallSeparatorPrefab;
             String objectName = "Wall_Separator";
-            Vector3 floorSize = _behavior._floorPrefab.GetComponentInChildren<MeshRenderer>().bounds.size;
+            Vector3 floorSize = _manager.levelStyle._floorPrefab.GetComponentInChildren<MeshRenderer>().bounds.size;
 
             float floorSpan = floorSize.x;
             float oneQuarterFloorSpan = floorSpan * 0.25f;
